% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count1bits.R
\name{25. Counting 1 bits}
\alias{25. Counting 1 bits}
\alias{count1bits}
\alias{count1bits.default}
\alias{count1bits.numeric}
\alias{count1bits.vli}
\title{Counting the Number of 1-Bits in vli Objects}
\usage{
count1bits(x)

\method{count1bits}{default}(x)

\method{count1bits}{numeric}(x)

\method{count1bits}{vli}(x)
}
\arguments{
\item{x}{object of class vli}
}
\value{
integer
}
\description{
Counting the number of 1-bits in the base 2 expression of vli (Very Large Integer) objects.
}
\examples{
x <- as.vli("69158247560284795612")
count1bits(x)
}
\author{
Javier Leiva Cuadrado
}
