% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divp2.R
\name{05. Efficent division by a power of 2}
\alias{05. Efficent division by a power of 2}
\alias{divp2}
\alias{divp2.default}
\alias{divp2.numeric}
\alias{divp2.vli}
\title{Efficient Division by a Power of 2}
\usage{
\method{divp2}{vli}(x, k)

\method{divp2}{default}(x, k)

\method{divp2}{numeric}(x, k)

\method{divp2}{vli}(x, k)
}
\arguments{
\item{x}{dividend; object of class vli or 32 bits integer}

\item{k}{exponent of the divisor (the divisor will be \code{2^k}); 32 bits integer}
}
\value{
object of class vli
}
\description{
\code{divp2} efficiently divides an object of class vli by a power of 2.
}
\details{
Given two integers \code{x} (vli or 32 bits integer) and \code{k} (32 bits integer), the function \code{divp2(x, k)} computes and returns \code{x/(2^k)} as an object of class vli.
}
\examples{
# Dividing a random 500 digits integer by 2^10 = 1024
x <- rvlidigits(500)
x
divp2(x, 10)
}
\author{
Javier Leiva Cuadrado
}
