% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{07. Factorial}
\alias{07. Factorial}
\alias{factvli}
\alias{factvli.default}
\alias{factvli.numeric}
\alias{factvli.vli}
\title{Factorial of a vli Object}
\usage{
factvli(n)

\method{factvli}{default}(n)

\method{factvli}{numeric}(n)

\method{factvli}{vli}(n)
}
\arguments{
\item{n}{object of class vli or 32 bits integer}
}
\value{
object of class vli
}
\description{
\code{factvli} computes and returns the factorial of a vli (Very Large Integers) object. Given a positive integer \code{n}, the factorial of \code{n}, \code{n!}, is defined as the product of all the positive integers from \code{1} to \code{n}.
}
\examples{
\dontrun{
n <- as.vli("420")
factvli(n)
}
}
\author{
Javier Leiva Cuadrado
}
