% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{09. Greatest common divisor}
\alias{09. Greatest common divisor}
\alias{gcd}
\alias{gcd.default}
\alias{gcd.numeric}
\alias{gcd.vli}
\title{Greatest Common Divisor for vli Objects}
\usage{
gcd(x, y)

\method{gcd}{default}(x, y)

\method{gcd}{numeric}(x, y)

\method{gcd}{vli}(x, y)
}
\arguments{
\item{x}{object of class vli or 32 bits integer}

\item{y}{object of class vli or 32 bits integer}
}
\value{
object of class vli
}
\description{
\code{gcd} computes and returns the greatest common divisor of two vli (Very Large Integers) objects.
}
\examples{
x <- as.vli("1225312091263347514461245")
y <- as.vli("357590484262521")
gcd(x, y)
}
\author{
Javier Leiva Cuadrado
}
