% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Legendre.R
\name{13. Legrendre's Formula}
\alias{13. Legrendre's Formula}
\alias{Legendre}
\alias{Legendre.default}
\alias{Legendre.numeric}
\alias{Legendre.vli}
\title{Legrendre's Formula for vli Objects}
\usage{
Legendre(n, p)

\method{Legendre}{default}(n, p)

\method{Legendre}{numeric}(n, p)

\method{Legendre}{vli}(n, p)
}
\arguments{
\item{n}{a positive integer; object of class vli or 32 bits integer}

\item{p}{a prime number; object of class vli or 32 bits integer}
}
\value{
object of class vli
}
\description{
Given a positive integer \code{n} and a prime \code{p}, the Legendre's Formula finds the largest integer \code{x} such that \code{p^x} divides the factorial of \code{n}, \code{n!}.
}
\examples{
p <- as.vli(577)
is.prime(p)
Legendre(12222, p)
}
\author{
Javier Leiva Cuadrado
}
